requirements.txt
yfinance
mplfinance
google-generativeai
GoogleNews
python-dotenv
pandas
requests
line-bot-sdk
trafilatura